requirements.txt
streamlit==1.28.0
google-generativeai==0.3.2
pandas==2.1.0
python-dotenv==1.0.0
pydantic==2.4.2
google-cloud-aiplatform==1.36.0 